referral_app/api/serializers.py
from rest_framework import serializers
from referral_app.models import Profile


class ProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profile
        fields = ('phone', 'ref_code')
