pyproject.toml
[tool.poetry]
name = "referral-service"
version = "0.1.0"
description = "Simple referral service"
authors = ["al-ov73 <klever1987@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
celery = "^5.4.0"
django-bootstrap5 = "^24.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
django-phone-field = "^1.8.1"
psycopg2-binary = "^2.9.9"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
djoser = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
